fix(ui): harden dashboard availability lifecycle
